Delay TT key calculation and probe.
Given that
https://github.com/GediminasMasaitis/4ku/commit/b0997e0054470dc27a22214bb848cf240ee0d6ba
surprisingly failed with -11 Elo, I wanted to see why that is,
because it superficially makes little sense that it could br weaker.
One possible reason is that get_pos() which calculates the TT key
is pretty slow and may not be all that much faster than the eval
itself, so getting the eval out of the TT is no real speedup.

That leads to this idea: avoid calculating the TT key and
doing the TT key lookup until the last possible moment.

Prefetching can in theory be done as a follow-up,
but naming __builtin_prefetch is too many bytes for 4ku.

ELO   | 3.85 +- 3.09 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=64MB
LLR   | 2.96 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 27648 W: 8043 L: 7737 D: 11868