Handle inverted wtime btime order. (#107)
Although the majority of UCI interfaces do the sensible thing
and just send wtime btime in that order, for whatever reason
ChessBase programs can start to send them in inverted order
in the middle of a game. Detect and handle this.

Not required for the 4k tournament as the command format
is defined in the rules.

No size impact.