Add ability to generate legal pawn moves.
Previous "optimizations" eliminated the default argument of
no promoted piece, but neglected to modify the callers,
rendering the engine unable to play chess. As we found that
capability rather useful, it seems worthwhile to spend a few
extra bytes to restore it.

Some minimal testing has been done to verify that this patch
is correct.