Fix kingsafety mask checking.
The bonus for a reasonable square never worked correctly because
we missed that the whole king branch already applies a king
position mask that is more restrictive.

I tested distinguishing between "reasonable" and "best" king
squares but the tuner tunes the difference in bonus to a really
small value, so that is not worth it.

f1 is also no longer considered a "good" square for the king
to be on - which is not coincidentally the only knowledge the
old version of this code could add.

-5 bytes

ELO   | 5.58 +- 5.30 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=64MB
LLR   | 2.96 (-2.94, 2.94) [-3.00, 1.00]
GAMES | N: 9968 W: 3096 L: 2936 D: 3936