Don't prune promotions and order them properly.
+1 byte

ELO   | 7.23 +- 5.24 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=64MB
LLR   | 2.97 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 9752 W: 2921 L: 2718 D: 4113

I observed that 4ku needs more than a 1 ply search
or qsearch to find the correct move in the following
position: 8/R7/4kP1p/8/8/4K3/6pN/8 b - - 0 1
Which itself was found by looking for very large
qsearch vs deep search inconsistencies.

The underlying cause is that KxP is not pruned,
but promoting to a queen is. That's unfortunate,
and fixing it seems to help.