Exploit early exit in attacked().
The return value of attacked() is an OR-ed bitmap of the pieces
doing the attacking. But there is no place in the code that
makes use of that, and one user explicitly has to convert the
result to boolean.

We can exploit early exit to make a lot of attack calculation
redundant. Unfortunately the is_attacked=true case isn't the
most common, so it's only ~1% speedup, but it does gain 10 bytes
at the cost of some readability.

Rename the function to make the return type more explicit.

STC

ELO   | 0.49 +- 2.35 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=64MB
LLR   | 2.95 (-2.94, 2.94) [-3.00, 1.00]
GAMES | N: 47728 W: 13652 L: 13585 D: 20491